publish_qc: merge the unit tests into "test_auto_processor.py".
